Move Navigation into a separate component
